+++ src/code/DbSqlLikeMem.Oracle/OracleTranslator.cs
@@ -1,6 +1,4 @@
-using System.Collections.Concurrent;
 using System.Linq.Expressions;
-using System.Reflection;
 using System.Text;
 
 namespace DbSqlLikeMem.Oracle;
@@ -10,402 +8,43 @@ namespace DbSqlLikeMem.Oracle;
 /// EN: Translates LINQ expressions into Oracle-compatible SQL statements.
 /// PT-br: Traduz expressões LINQ para instruções SQL compatíveis com Oracle.
 /// </summary>
-public class OracleTranslator : ExpressionVisitor
+public class OracleTranslator : TranslatorBase<OracleTranslator>
 {
-    private const int TranslationCacheSoftLimit = 256;
-    private static readonly ConcurrentDictionary<string, string> TranslationCache = new(StringComparer.Ordinal);
-    private StringBuilder _sb = new();
-    private readonly List<object> _values = [];
-    private string? _table;
-    private string? _projection;
-    private string? _whereClause;
-    private string? _orderByClause;
-    private int? _offset;
-    private int? _limit;
+    /// <inheritdoc />
+    protected override Type QueryableTypeDefinition => typeof(OracleQueryable<>);
 
     /// <summary>
-    /// EN: Translates a LINQ expression into SQL and parameters.
-    /// PT-br: Traduz uma expressão LINQ em SQL e parâmetros.
+    /// EN: Appends Oracle-specific pagination using OFFSET ... ROWS FETCH NEXT/FIRST ... ROWS ONLY.
+    /// PT-br: Adiciona paginacao especifica do Oracle usando OFFSET ... ROWS FETCH NEXT/FIRST ... ROWS ONLY.
     /// </summary>
-    public TranslationResult Translate(Expression expression)
+    protected override void AppendPagination(StringBuilder sb, int? offset, int? limit)
     {
-        var cacheKey = BuildTranslationCacheKey(expression);
-        if (TranslationCache.TryGetValue(cacheKey, out var cachedSql))
-            return new TranslationResult(cachedSql, BuildParameters(CollectParameters(expression)));
-
-        _sb.Clear();
-        _values.Clear();
-        _table = null;
-        _whereClause = null;
-        _orderByClause = null;
-        _offset = null;
-        _limit = null;
-
-        Visit(expression);
-
-        _sb.Append("SELECT ");
-        _sb.Append(string.IsNullOrWhiteSpace(_projection) ? "*" : _projection);
-        _sb.Append(" FROM ").Append(_table);
-        if (!string.IsNullOrWhiteSpace(_whereClause))
-            _sb.Append(" WHERE ").Append(_whereClause);
-        if (!string.IsNullOrWhiteSpace(_orderByClause))
-            _sb.Append(" ORDER BY ").Append(_orderByClause);
-        if (_offset.HasValue)
-            _sb.Append(" OFFSET ").Append(_offset.Value).Append(" ROWS");
-        if (_limit.HasValue)
+        if (offset.HasValue)
+            sb.Append(" OFFSET ").Append(offset.Value).Append(" ROWS");
+        if (limit.HasValue)
         {
-            if (_offset.HasValue)
-                _sb.Append(" FETCH NEXT ").Append(_limit.Value).Append(" ROWS ONLY");
+            if (offset.HasValue)
+                sb.Append(" FETCH NEXT ").Append(limit.Value).Append(" ROWS ONLY");
             else
-                _sb.Append(" FETCH FIRST ").Append(_limit.Value).Append(" ROWS ONLY");
+                sb.Append(" FETCH FIRST ").Append(limit.Value).Append(" ROWS ONLY");
         }
-
-        var sql = _sb.ToString();
-        CacheTranslation(cacheKey, sql);
-        return new TranslationResult(sql, BuildParameters(_values));
-    }
-
-#pragma warning disable CA1859 // Use concrete types when possible for improved performance
-    private static object BuildParameters(List<object> vals)
-#pragma warning restore CA1859 // Use concrete types when possible for improved performance
-    {
-        var dict = new Dictionary<string, object>();
-        for (int i = 0; i < vals.Count; i++)
-            dict[$"p{i}"] = vals[i];
-        return dict;
-    }
-
-    private static string BuildTranslationCacheKey(Expression expression)
-    {
-        var tableName = TryExtractTableName(expression) ?? string.Empty;
-        var expressionText = expression.ToString();
-        return string.Concat(tableName, "|", expressionText);
-    }
-
-    private static void CacheTranslation(string cacheKey, string sql)
-    {
-        if (TranslationCache.Count >= TranslationCacheSoftLimit)
-            TranslationCache.Clear();
-
-        TranslationCache[cacheKey] = sql;
-    }
-
-    private static List<object> CollectParameters(Expression expression)
-    {
-        var collector = new OracleTranslationParameterCollector();
-        collector.Visit(expression);
-        return collector.Values;
-    }
-
-#pragma warning disable CS8605 // Unboxing a possibly null value.
-    /// <summary>
-    /// EN: Represents Visit Method Call.
-    /// PT-br: Representa Visit Method Call.
-    /// </summary>
-    protected override Expression VisitMethodCall(MethodCallExpression node)
-    {
-        ArgumentNullExceptionCompatible.ThrowIfNull(node, nameof(node));
-        var method = node.Method.Name;
-        if (method == "Where")
-        {
-            Visit(node.Arguments[0]);
-            var lambda = (LambdaExpression)StripQuotes(node.Arguments[1]);
-
-            var whereBuilder = new StringBuilder();
-            var oldSb = _sb;
-            _sb = whereBuilder;
-            Visit(lambda.Body);
-            _whereClause = whereBuilder.ToString();
-            _sb = oldSb;
-            return node;
-        }
-        if (method.StartsWith("OrderBy", StringComparison.Ordinal)
-            || method.StartsWith("ThenBy", StringComparison.Ordinal))
-        {
-            Visit(node.Arguments[0]);
-            var lambda = (LambdaExpression)StripQuotes(node.Arguments[1]);
-            var member = (MemberExpression)lambda.Body;
-
-            if (_orderByClause == null)
-                _orderByClause = member.Member.Name;
-            else
-                _orderByClause += ", " + member.Member.Name;
-
-            if (method.EndsWith("Descending", StringComparison.OrdinalIgnoreCase))
-                _orderByClause += " DESC";
-
-            return node;
-        }
-        if (method == "Skip")
-        {
-            Visit(node.Arguments[0]);
-            _offset = (int)((ConstantExpression)node.Arguments[1]).Value;
-            return node;
-        }
-        if (method == "Take")
-        {
-            Visit(node.Arguments[0]);
-            _limit = (int)((ConstantExpression)node.Arguments[1]).Value;
-            return node;
-        }
-        if (method == "Count")
-        {
-            Visit(node.Arguments[0]); // garante _table
-            _projection = "COUNT(*)";
-            return node;
-        }
-        if (method == "Select")
-        {
-            Visit(node.Arguments[0]); // Visita a fonte
-
-            var lambda = (LambdaExpression)StripQuotes(node.Arguments[1]);
-            _projection = BuildProjection(lambda);
-            return node;
-        }
-
-        // SqlFunctions: Contains
-        if (node.Method.DeclaringType == typeof(SqlFunctions))
-        {
-            switch (method)
-            {
-                case nameof(SqlFunctions.Contains):
-                    _sb.Append("CONTAINS(");
-                    Visit(node.Arguments[0]);
-                    _sb.Append(", ");
-                    Visit(node.Arguments[1]);
-                    _sb.Append(')');
-                    return node;
-                default:
-                    return base.VisitMethodCall(node);
-            }
-        }
-
-        // raiz: Query<TEntity>
-        return base.VisitMethodCall(node);
-    }
-#pragma warning restore CS8605 // Unboxing a possibly null value.
-
-    /// <summary>
-    /// EN: Represents Visit Constant.
-    /// PT-br: Representa Visit Constant.
-    /// </summary>
-    protected override Expression VisitConstant(ConstantExpression node)
-    {
-        ArgumentNullExceptionCompatible.ThrowIfNull(node, nameof(node));
-
-        // 1) Prioridade: se for OracleQueryable<>, usa TableName (é o que você passou em AsQueryable<T>("users"))
-        if (node.Value is not null)
-        {
-            var t = node.Value.GetType();
-            if (t.IsGenericType && t.GetGenericTypeDefinition() == typeof(OracleQueryable<>))
-            {
-                var prop = t.GetProperty("TableName", BindingFlags.Public | BindingFlags.Instance);
-                if (prop?.GetValue(node.Value) is string tn && !string.IsNullOrWhiteSpace(tn))
-                {
-                    _table = tn;
-                    return node;
-                }
-            }
-        }
-
-        // 2) Fallback: qualquer IQueryable que não seja o seu root custom (usa nome do tipo)
-        if (node.Value is IQueryable q)
-        {
-            _table = q.ElementType.Name;
-            return node;
-        }
-
-        // 3) Constante normal vira parâmetro
-        if (node.Value != null)
-        {
-            var idx = _values.Count;
-            _values.Add(node.Value);
-            _sb.Append($"@p{idx}");
-        }
-
-        return node;
     }
 
     /// <summary>
-    /// EN: Represents Visit Binary.
-    /// PT-br: Representa Visit Binary.
+    /// EN: Translates SqlFunctions.Contains into CONTAINS().
+    /// PT-br: Traduz SqlFunctions.Contains para CONTAINS().
     /// </summary>
-    protected override Expression VisitBinary(BinaryExpression node)
+    protected override bool TryTranslateSqlFunction(StringBuilder sb, string method, MethodCallExpression node)
     {
-        ArgumentNullExceptionCompatible.ThrowIfNull(node, nameof(node));
-        _sb.Append('(');
-        Visit(node.Left);
-        switch (node.NodeType)
-        {
-            case ExpressionType.Equal: _sb.Append(" = "); break;
-            case ExpressionType.GreaterThan: _sb.Append(" > "); break;
-            case ExpressionType.LessThan: _sb.Append(" < "); break;
-            case ExpressionType.AndAlso: _sb.Append(SqlConst.AND_SPACED); break;
-            case ExpressionType.OrElse: _sb.Append(" OR "); break;
-            default: _sb.Append(' '); break;
-        }
-        Visit(node.Right);
-        _sb.Append(')');
-        return node;
-    }
+        if (method != nameof(SqlFunctions.Contains))
+            return false;
 
-    /// <summary>
-    /// EN: Represents Visit Member.
-    /// PT-br: Representa Visit Member.
-    /// </summary>
-    protected override Expression VisitMember(MemberExpression node)
-    {
-        ArgumentNullExceptionCompatible.ThrowIfNull(node, nameof(node));
-        if (node.Expression != null
-            && node.Expression.NodeType == ExpressionType.Parameter)
-        {
-            _sb.Append(node.Member.Name);
-            return node;
-        }
-        // acesso a constante (captured variable)
-        var value = Expression.Lambda(node).Compile().DynamicInvoke();
-        var idx = _values.Count;
-        if (value != null)
-        {
-            _values.Add(value);
-            _sb.Append($"@p{idx}");
-        }
-        return node;
-    }
-
-    private static Expression StripQuotes(Expression e)
-    {
-        while (e.NodeType == ExpressionType.Quote)
-            e = ((UnaryExpression)e).Operand;
-        return e;
-    }
-
-    private static string BuildProjection(LambdaExpression lambda)
-    {
-        if (lambda.Body is NewExpression nex)
-        {
-            // exemplo: u => new { u.X, u.Y }
-            return string.Join(", ",
-                nex.Arguments.Zip(nex.Members ?? new List<MemberInfo>().AsReadOnly(), (arg, member) =>
-                {
-                    if (arg is MemberExpression me)
-                    {
-                        var name = me.Member.Name;
-                        return name == member.Name ? name : $"{name} AS {member.Name}";
-                    }
-                    return member.Name;
-                }));
-        }
-
-        if (lambda.Body is MemberExpression mex)
-        {
-            // exemplo: u => u.X
-            return mex.Member.Name;
-        }
-
-        return "*"; // fallback
-    }
-
-    private static string? TryExtractTableName(Expression expression)
-    {
-        switch (expression)
-        {
-            case ConstantExpression constant:
-                {
-                    if (constant.Value is null)
-                        return null;
-
-                    var valueType = constant.Value.GetType();
-                    if (valueType.IsGenericType && valueType.GetGenericTypeDefinition() == typeof(OracleQueryable<>))
-                    {
-                        var property = valueType.GetProperty(nameof(OracleQueryable<object>.TableName), BindingFlags.Instance | BindingFlags.Public);
-                        if (property?.GetValue(constant.Value) is string tableName && !string.IsNullOrWhiteSpace(tableName))
-                            return tableName;
-                    }
-
-                    if (constant.Value is IQueryable queryable)
-                        return queryable.ElementType.Name;
-
-                    return null;
-                }
-
-            case MethodCallExpression methodCall when methodCall.Arguments.Count > 0:
-                return TryExtractTableName(methodCall.Arguments[0]);
-
-            default:
-                return null;
-        }
-    }
-
-    private sealed class OracleTranslationParameterCollector : ExpressionVisitor
-    {
-        internal List<object> Values { get; } = [];
-
-        protected override Expression VisitMethodCall(MethodCallExpression node)
-        {
-            ArgumentNullExceptionCompatible.ThrowIfNull(node, nameof(node));
-
-            if (node.Method.Name is "Where" or "OrderBy" or "OrderByDescending" or "ThenBy" or "ThenByDescending" or "Select")
-            {
-                Visit(node.Arguments[0]);
-                Visit(StripQuotes(node.Arguments[1]));
-                return node;
-            }
-
-            if (node.Method.Name is "Skip" or "Take")
-            {
-                Visit(node.Arguments[0]);
-                Visit(node.Arguments[1]);
-                return node;
-            }
-
-            if (node.Method.Name == "Count")
-            {
-                Visit(node.Arguments[0]);
-                return node;
-            }
-
-            if (node.Method.DeclaringType == typeof(SqlFunctions))
-            {
-                foreach (var arg in node.Arguments)
-                    Visit(arg);
-                return node;
-            }
-
-            return base.VisitMethodCall(node);
-        }
-
-        protected override Expression VisitConstant(ConstantExpression node)
-        {
-            ArgumentNullExceptionCompatible.ThrowIfNull(node, nameof(node));
-
-            if (node.Value is null)
-                return node;
-
-            var valueType = node.Value.GetType();
-            if (valueType.IsGenericType && valueType.GetGenericTypeDefinition() == typeof(OracleQueryable<>))
-                return node;
-
-            if (node.Value is IQueryable)
-                return node;
-
-            Values.Add(node.Value);
-            return node;
-        }
-
-        protected override Expression VisitMember(MemberExpression node)
-        {
-            ArgumentNullExceptionCompatible.ThrowIfNull(node, nameof(node));
-
-            if (node.Expression is not null && node.Expression.NodeType == ExpressionType.Parameter)
-                return node;
-
-            var value = Expression.Lambda(node).Compile().DynamicInvoke();
-            if (value is not null)
-                Values.Add(value);
-
-            return node;
-        }
+        sb.Append("CONTAINS(");
+        Visit(node.Arguments[0]);
+        sb.Append(", ");
+        Visit(node.Arguments[1]);
+        sb.Append(')');
+        return true;
     }
 }
 #pragma warning restore CA1305 // Specify IFormatProvider
